requirements.txt
streamlit 
torch 
torchvision 
pandas 
matplotlib 
pillow
opencv-python 